requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-embed-video==1.4.7
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
